+++ android/app/src/main/kotlin/com/mb28/monoP/ExifDart.kt
@@ -1,4 +1,4 @@
-package com.example.memories_photos
+package com.mb28.monoP
 
 import androidx.exifinterface.media.ExifInterface as EX