+++ lib/Scripts/android_helper.dart
@@ -1,7 +1,9 @@
-// ignore_for_file: empty_catches
+// ignore_for_file: constant_identifier_names, empty_catches
 
 import 'dart:io';
+import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:memories_photos/Popups/toast.dart';
 
 class AndroidHelper {
   static const channel = MethodChannel("mb28.monoP.helper/helper_channel");
@@ -32,6 +34,30 @@ class AndroidHelper {
 
       await channel.invokeMethod("openWithMaps", {"lat": latPos.toString(), "long": longPos.toString()});
     } on PlatformException {}
+  }
+
+  static void startActivityClass(String package, String classN) async {
+    try {
+      await channel.invokeMethod("startActivityClass", {"class": classN, "package": package});
+    } on PlatformException{}
+  }
+
+  static const String MODE_PORTRAIT = "com.sec.android.app.camera.shootingmode.livefocus";
+  static const String MODE_PRO = "com.sec.android.app.camera.shootingmode.pro";
+  static const String MODE_PRO_VIDEO = "com.sec.android.app.camera.shootingmode.provideo";
+  static void startSamsungCameraWidget(BuildContext context, String mode) async {
+    try {
+      await channel.invokeMethod("startSamsungCameraWidget", {"mode": mode});
+    } on PlatformException {
+      showStyledToast("Samsung Camera (com.sec.android.app.camera) is not installed!", context, duration: 5);
+    }
+  }
+  static void startSamsungCameraWidgetSettings(BuildContext context) async {
+    try {
+      await channel.invokeMethod("startSamsungCameraWidgetSettings");
+    } on PlatformException catch (e) {
+      showStyledToast("Samsung Camera (com.sec.android.app.camera) is not installed!\n\n$e", context, duration: 5);
+    }
   }
 
 }